+++ src/components/Eamplemap.jsx
@@ -1,17 +1,26 @@
 import React from 'react'
-import { GoogleMap, useJsApiLoader,Marker } from '@react-google-maps/api';
+import Container from 'react-bootstrap/esm/Container';
+import { Row } from 'react-bootstrap';
+import TopNav from '../components/TopNav';
+import Footer from '../components/Footer';
+import Button from 'react-bootstrap/Button';
+//import './Pages.css
+
+
+import { GoogleMap, useJsApiLoader,Marker,useLoadScript } from '@react-google-maps/api';
+import { useState } from 'react';
+import { Link, useLocation } from 'react-router-dom';
 
 const containerStyle = {
-  width: '100%',
+  width: '100px',
   height: '800px'
 };
 
-const center = {
-  lat: 7.285039899999999,
-  lng: 80.6169968,
-};
 
-function MyComponent() {
+function Eamplemap (props) {
+
+  const pathForPage=useLocation().state?.data;
+  console.log(pathForPage);
   const { isLoaded } = useJsApiLoader({
     id: 'my-project-cs304-371916',
     googleMapsApiKey: "AIzaSyD0Zw3uHJdRndDSN5OnCa1aflSbltWKtzg"
@@ -19,10 +28,17 @@ function MyComponent() {
 
   const [map, setMap] = React.useState(null)
 
+  const [coordinates, setCoordinates] = useState({lat:6.927079, lng:79.861244});
+
+  const handleDragEnd =(e) => {
+    setCoordinates({lat:e.latLng.lat(), lng:e.latLng.lng()});
+    console.log(coordinates);
+  }
+
   const onLoad = React.useCallback(function callback(map) {
     
-    const bounds = new window.google.maps.LatLngBounds(center);
-    map.fitBounds(bounds);
+    const bounds = new window.google.maps.LatLngBounds(coordinates);
+   map.fitBounds(bounds);
 
     setMap(map)
   }, [])
@@ -31,21 +47,50 @@ function MyComponent() {
     setMap(null)
   }, [])
 
+  const success="location saved";
+
   return isLoaded ? (
-      <GoogleMap
+    <><div><TopNav></TopNav></div>
+         <Container>
+          <Row>
+            {pathForPage===1?
+          <Link to="/signing"  state={{data: coordinates,success}} className="Link">
+            <Button className="btn btn-success" > save location</Button></Link>
+              :
+              <Link to="/driverProfile"  state={{data: coordinates,success}} className="Link">
+            <Button className="btn btn-success" > save location</Button>
+            <Button className="btn btn-danger"  style={{marginLeft:"10px"}}> Cancel
+      </Button>
+            </Link>
+          }
+      
+      </Row>
+   
+      
+       <Row>
+        <div className="card col-md-10 ">
+               
+               <div className="card-body">
+               <GoogleMap
         mapContainerStyle={containerStyle}
-        center={center}
-        zoom={10}
+        center={coordinates}
+        zoom={9}
         onLoad={onLoad}
         onUnmount={onUnmount}
         
       >
         { <Marker
-      
-    />/* Child components, such as markers, info windows, etc. */ }
-        <></>
+          position={coordinates}
+          onDragEnd={handleDragEnd}
+          draggable={true}
+          />}
       </GoogleMap>
-  ) : <></>
+        </div>
+    </div>
+    </Row>
+    </Container>
+      </>
+  ): <></>
 }
 
-export default React.memo(MyComponent)
+export default  React.memo(Eamplemap)
