+++ src/pages/driverPages/OrderOfShedule.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
-import DriverNavbar from '../../components/DriverNavbar';
+import DriverNavbar from '../../components/Driver/DriverNavbar';
 import UserMap from '../../components/UserMap';
 import './DriverCss.css'
