+++ src/pages/globals.css
@@ -17,19 +17,20 @@ a {
 
 .places-container {
   position: absolute;
-  top: 10px;
+  top: 50px;
   left: 50%;
   transform: translateX(-50%);
   z-index: 10;
-  width: 300px;
+  width: 500px;
 }
 
 .map-container {
-  width: 800px;
+  width: 100%;
   height: 800px;
 }
 
 .combobox-input {
   width: 100%;
-  padding: 0.5rem;
+  padding: 1rem;
+ 
 }
